bot/requirements.txt
aiogram==2.23.1
aio-pika==8.2.5
python-dotenv==0.21.1